Add main notes binary kicking of the notetaker
The notes binary outputs general usage and runs subcommands with the
notes-prefix.

For now the notes binary allows only a specific set of commands for
simplicity.

It doesn't care about notes paths configuration or anything like that.

It does not check subcommands syntax, it just passes on the options and
arguments passed after a recognised subcommand.